clean up debian/rules
Overriding dh_auto_install isn't necessary if we build more than one binary package. And explicitly setting the compression to xz is not needed too since nowadays xz is set by default (this would be necessary for Ubuntu 12.04, but because of missing build dependencies this package doesn't build on that distribution anyway).